+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## [unreleased]
+- Remove any chr prefix from input chromosome contig name
+
 ## [0.9]
 - Blacklist (excluded regions list) also hides probes